Move files + correct test
