+++ packages/lib/checkRateLimitAndThrow.test.ts
@@ -1,27 +1,47 @@
-import {default as rateLimit, checkRateLimitAndThrowError } from './rateLimit';
+import { checkRateLimitAndThrowError } from './checkRateLimitAndThrow';
+import { rateLimiter  } from './rateLimit';
+import type { RatelimitResponse } from './rateLimit';
 import { describe, expect, it,vi } from "vitest";
 
 
+vi.mock("./rateLimit",() => {
+  return ({
+    rateLimiter: vi.fn(),
+  });
+})
 
 describe('checkRateLimitAndThrowError', () => {
   it('should throw an error if rate limit is exceeded', async () => {
     process.env.UPSTASH_REDIS_REST_URL="mockUrl"
     process.env.UPSTASH_REDIS_REST_TOKEN="mockToken"
-    const rateLimiterMock = vi.fn().mockImplementation(rateLimit);
 
+    vi.mocked(rateLimiter).mockReturnValue(() => {
+      return {
+        limit:10,
+        remaining:-1,
+        reset: Date.now() + 10000,
+      } as RatelimitResponse
+    })
+
+  
     const identifier = 'test-identifier';
     const rateLimitingType = 'core';
 
     await expect(
       checkRateLimitAndThrowError({ rateLimitingType, identifier })
-    ).rejects.toThrowErrorMatchingSnapshot();
+    ).rejects.toThrow();
   });
 
   it('should not throw an error if rate limit is not exceeded', async () => {
-    const rateLimiterMock = vi.fn().mockReturnValue({
-      remaining: 1,
-      reset: Date.now() + 10000,
-    });
+    process.env.UPSTASH_REDIS_REST_URL="mockUrl"
+    process.env.UPSTASH_REDIS_REST_TOKEN="mockToken"
+    vi.mocked(rateLimiter).mockReturnValue(() => {
+      return {
+        limit:10,
+        remaining:5,
+        reset: Date.now() + 10000,
+      } as RatelimitResponse
+    })
 
     const identifier = 'test-identifier';
     const rateLimitingType = 'core';
